+++ hidden-api/src/main/java/android/content/pm/IPackageManager.java
@@ -10,7 +10,7 @@ public interface IPackageManager extends IInterface {
 
     void setSplashScreenTheme(String packageName, String themeName, int userId) throws RemoteException;
 
-    ParceledListSlice queryIntentActivities(Intent intent, String resolvedType, int flags, int userId) throws RemoteException;
+    ParceledListSlice queryIntentActivities(Intent intent, String resolvedType, long flags, int userId) throws RemoteException;
 
     abstract class Stub extends Binder implements IPackageManager {
         public static IPackageManager asInterface(IBinder binder) {
